Changing frontend API endpoints and moving to correct files. Changing frontend test js file route mappings and moving to correct test files.
