+++ frontend/react-app/src/api/authInfoApi.js
